requirements.txt
langchain==0.1.6
langchain-community==0.0.19
langchain-openai
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
sentence-transformers>=2.5.0
huggingface-hub
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0
colorama>=0.4.6
python-dotenv>=1.0.0
streamlit>=1.30.0
openai>=1.0.0
